enhances for rpc/literal support.  for #118.
git-svn-id: http://dev.ctor.org/svn/soap4r/trunk@1581 23c4dcf6-64ea-0310-84d7-d424c7495ec0

+++ test/wsdl/soap/wsdl2ruby/section/test_section.rb
@@ -11,7 +11,7 @@ class TestSection < Test::Unit::TestCase
   end
 
   def teardown
-    File.unlink(pathname("mysample.rb"))
+    File.unlink(pathname("mysample.rb")) unless $DEBUG
   end
 
   def test_classdef
